feat(cli): cleanup code and split functionality
